refactor live stage controls
